feat: add SetPositionAfter that seeks from a byte index
SetPosition works fine, but it needs to seek through a large part of the
file to get to the correct location. This can be expensive, especially
when the file is read in chunks on demand from the network (it isn't
fully buffered in advance).

Some applications (in my case, Spotify) have a seek index, with which
the application can know that a given audio sample is somewhere after a
particular byte position. Starting the seek at this index greatly speeds
up seeking.